+++ src/plugins/test-utils/schema-objects.ts
@@ -24,18 +24,29 @@ appendToArray(allChars, someEmojisArr);
 
 export function randomStringWithSpecialChars(length: number) {
     let text = '';
-    for (let i = 0; i < length; i++) {
-        if(i === 0){
+
+    while (text.length < length) {
+        if (text.length === 0) {
             /**
              * TODO foundationdb does not work correctly when an index string starts
              * with an emoji. This can likely be fixed by upgrading foundationdb to the
              * latest version.
              */
             text += baseChars[Math.floor(Math.random() * baseChars.length)];
-        }else {
+        } else {
             text += allChars[Math.floor(Math.random() * allChars.length)];
         }
     }
+
+    /**
+     * Because emojis can have a string.length of 2,
+     * we can sometimes end up with strings that are longer
+     * than the provided length. In that cases we have to rerun.
+     */
+    if (text.length > length) {
+        return randomStringWithSpecialChars(length);
+    }
+
     return text;
 }